pyproject.toml
[tool.poetry]
name = "military-tracker-bot"
version = "0.1.0"
description = "Telegram bot for military personnel tracking"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.7"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"